Clean up and reorganize gwas-related files
