+++ snipar/gwas.py
@@ -6,7 +6,7 @@ from pysnptools.snpreader import Bed
 from scipy.stats import chi2
 from math import log10
 import snipar.read as read
-import snipar.lmm as lmm
+import snipar.slmm as slmm
 from snipar.gtarray import impute_missing, impute_unrel_par_gts
 from snipar.utilities import *
 from numba import njit, prange
@@ -15,10 +15,10 @@ from typing import List, Union, Dict, Hashable
 from multiprocessing import Pool, RawArray
 from functools import partial
 import ctypes
-import logging
+# import logging
 
 
-logger = logging.getLogger(__name__)
+# logger = logging.getLogger(__name__)
 
 
 def find_common_ind_ids(obsfiles, impfiles, pheno_ids, from_chr=None, covar=None, keep=None, include_unrel=False):
@@ -40,7 +40,6 @@ def find_common_ind_ids(obsfiles, impfiles, pheno_ids, from_chr=None, covar=None
             df = df.reset_index()
             ids = df['index'].values
             fam_labels = df[f'fam_labels'].values
-        # logger.info(f'Found {len(ids)} common ids.')
         return ids, fam_labels
     first = True
     j = from_chr
@@ -60,10 +59,6 @@ def find_common_ind_ids(obsfiles, impfiles, pheno_ids, from_chr=None, covar=None
             df = df.merge(df_, how='inner', left_index=True, right_index=True)
     if len(df.index) == 0:
         raise ValueError('No commond ids among chromosome files.')
-    # for j in range(1, 22):
-    #     if not df[f'fam_labels_{j}'].equals(df[f'fam_labels_{j + 1}']):
-    #         raise ValueError(
-    #             'fam_labels are not consistent across chromosome files.')
     if covar is not None:
         df_covar = pd.DataFrame(index=covar.ids)
         df = df.merge(df_covar, how='inner', left_index=True, right_index=True)
@@ -77,44 +72,6 @@ def find_common_ind_ids(obsfiles, impfiles, pheno_ids, from_chr=None, covar=None
     fam_labels = df[f'fam_labels_{j}'].values
     return ids, fam_labels
 
-
-def transform_phenotype(inv_root, y, fam_indices, null_mean = None):
-    """
-    Transform phenotype based on inverse square root of phenotypic covariance matrix.
-    If the null model included covariates, the fitted mean is removed rather than the overall mean
-    """
-    # Mean normalise phenotype
-    if null_mean is None:
-        y = y - np.mean(y)
-    else:
-        y = y - null_mean
-    # Transform by family
-    for fam in fam_indices.keys():
-        famsize = fam_indices[fam].shape[0]
-        if famsize == 1:
-            y[fam_indices[fam]] = inv_root[1] * y[fam_indices[fam]]
-        else:
-            y[fam_indices[fam]] = inv_root[famsize].dot(y[fam_indices[fam]])
-    return y
-
-@njit(parallel=True)
-def fit_models(y,G):
-    alpha = np.zeros((G.shape[2],G.shape[1]),dtype=np.float_)
-    alpha_cov = np.zeros((G.shape[2],G.shape[1],G.shape[1]),dtype=np.float_)
-    for i in prange(G.shape[2]):
-        not_na = np.sum(np.isnan(G[:,:,i]),axis=1)==0
-        xtx = G[not_na,:,i].T @ (G[not_na,:,i])
-        xty = G[not_na,:,i].T @ y[not_na]
-        alpha[i,:] = np.linalg.solve(xtx,xty)
-        alpha_cov[i,:,:] = np.linalg.inv(xtx)
-    return alpha, alpha_cov
-
-@njit(parallel=True)
-def compute_ses(alpha_cov):
-    alpha_ses = np.zeros((alpha_cov.shape[0],alpha_cov.shape[1]),dtype=np.float_)
-    for i in prange(alpha_cov.shape[0]):
-        alpha_ses[i,:] = np.sqrt(np.diag(alpha_cov[i,:,:]))
-    return alpha_ses
 
 def write_output(chrom, snp_ids, pos, alleles, outfile, parsum, sib, sib_diff, alpha, alpha_ses, alpha_cov, sigmas, freqs, robust=False, standard_gwas=False):
     """
@@ -130,9 +87,6 @@ def write_output(chrom, snp_ids, pos, alleles, outfile, parsum, sib, sib_diff, a
         if sib:
             X_length += 1
             outcols.append('sib')
-        # if sib_diff:
-        #     X_length += 1
-        #     outcols.append('fam')
         if parsum:
             X_length += 1
             outcols.append('avg_NTC')
@@ -174,10 +128,6 @@ def write_txt_output(chrom, snp_ids, pos, alleles, outfile, parsum, sib, sib_dif
         outstack = [outbim]
         outstack.append(outarray_effect(alpha[:,0],np.sqrt(alpha_cov[:,0,0]),freqs,vy))
         header += [effects[i]+'_N',effects[i]+'_Beta',effects[i]+'_SE',effects[i]+'_Z',effects[i]+'_log10_P']
-        # outstack.append(np.round(alpha_corr_out,6))
-        # if sib_diff:
-        #     outstack.append(np.round(alpha[:,1], decimals=6))
-        #     header += ['fixed'+'_Beta']
         # Output array
         outarray = np.row_stack((np.array(header),np.column_stack(outstack)))
         print('Writing text output to '+outfile)
@@ -261,8 +211,6 @@ def split_batches(nsnp, niid, nbytes, num_cpus, parsum=False, sib=False):
     while int(nsnp / n_tasks) * nbytes > max_nbytes / 4 or f(nsnp, niid, n_tasks) > max_nbytes / 4:
         n_tasks += num_cpus
     return np.array_split(np.arange(nsnp), n_tasks)
-    # return np.array_split(np.arange(nsnp), 1)
-    # return np.array_split(np.arange(nsnp), nsnp)
 
 
 _var_dict = {}
@@ -282,43 +230,10 @@ def _init_worker(y_, n_vararr_, varcomps_, covar_, covar_shape, **kwargs):
     # _var_dict['ped_shape'] = ped_shape
 
 
-# def process_batch(y, pedigree, tau, sigma2, snp_ids=None, bedfile=None, bgenfile=None, par_gts_f=None, parsum=False,
-#                   fit_sib=False, max_missing=5, min_maf=0.01, verbose=False, print_sample_info=False):
-#     ####### Construct family based genotype matrix #######
-#     G = read.get_gts_matrix(ped=pedigree, bedfile=bedfile, bgenfile=bgenfile, par_gts_f=par_gts_f, snp_ids=snp_ids, 
-#                                 ids=y.ids, parsum=parsum, sib=fit_sib, verbose=verbose, print_sample_info=print_sample_info)
-#     G.compute_freqs()
-#     #### Filter SNPs ####
-#     if verbose:
-#         print('Filtering based on MAF')
-#     G.filter_maf(min_maf)
-#     if verbose:
-#         print('Filtering based on missingness')
-#     G.filter_missingness(max_missing)
-#     if verbose:
-#         print(str(G.shape[2])+' SNPs that pass filters')
-#     #### Match phenotype ####
-#     y.filter_ids(G.ids)
-#     if G.ids.shape[0] > y.ids.shape[0]:
-#         G.filter_ids(y.ids)
-#     ##### Transform genotypes ######
-#     if verbose:
-#         print('Transforming genotypes')
-#     null_model = lmm.model(y.gts[:,0], np.ones((y.shape[0], 1)), y.fams)
-#     L = null_model.sigma_inv_root(tau, sigma2)
-#     G.diagonalise(L)
-#     ### Fit models for SNPs ###
-#     if verbose:
-#         print('Estimating SNP effects')
-#     alpha, alpha_cov = fit_models(np.array(y.gts[:,0],dtype=np.float32),G.gts)
-#     alpha_ses = compute_ses(alpha_cov)
-#     return G.freqs, G.sid, alpha, alpha_cov, alpha_ses
-
-
 def process_batch(snp_ids, ped,  n_vararr, imp_fams, pheno_ids=None, bedfile=None, bgenfile=None, par_gts_f=None,
                   parsum=False, fit_sib=False, sib_diff=False, max_missing=5, min_maf=0.01, verbose=False, 
                   print_sample_info=False, impute_unrel=False, unrelated_inds=None, standard_gwas=False,
-                  ignore_na_fams=False, ignore_na_rows=False, robust=False, cond_gaussian=False,
+                  robust=False, cond_gaussian=False,
                   add_jitter=False):
     y = np.frombuffer(_var_dict['y_'], dtype='float')
     varcomps = _var_dict['varcomps_']
@@ -349,71 +264,47 @@ def process_batch(snp_ids, ped,  n_vararr, imp_fams, pheno_ids=None, bedfile=Non
     G.compute_freqs()
     # impute parental genotypes of unrelated individuals
     if (impute_unrel and cond_gaussian) and not sib_diff:
-        # unrelated_inds = np.where(np.array([G.par_status[i, 0] == 0 and G.par_status[i, 1] == 0 for i in range(len(G.par_status))]))[0]
-        # true_par = G.gts[unrelated_inds, 1, :]
-        # par_ids = []
-        # for i in ped:
-        #     if i[4] == 'True' and i[5] == 'True':
-        #         if i[2] in pheno_ids:
-        #             par_ids.append(i[2])
-        #         if i[3] in pheno_ids:
-        #             par_ids.append(i[3])
-        # par_ids = np.unique(par_ids)
-        G = impute_unrel_par_gts(G, sib=fit_sib, parsum=parsum, ped=ped, unrelated_inds=unrelated_inds, grm=varcomp_arr_lst[0], ) #true_par=true_par, par_ids=par_ids)
-        # return G
+        G = impute_unrel_par_gts(G, sib=fit_sib, parsum=parsum, ped=ped, unrelated_inds=unrelated_inds, grm=varcomp_arr_lst[0])
     elif impute_unrel and not sib_diff:
         G = impute_unrel_par_gts(G, sib=fit_sib, parsum=parsum, ped=ped)
-    else:
-        logger.info('Not imputing unrelated individuals parental geno.')
     #### Filter SNPs ####
     if verbose:
-        logger.info('Filtering based on MAF')
+        # logger.info('Filtering based on MAF')
+        print('Filtering based on MAF')
     G.filter_maf(min_maf)
     if verbose:
-        logger.info('Filtering based on missingness')
+        # logger.info('Filtering based on missingness')
+        print('Filtering based on missingness')
     G.filter_missingness(max_missing)
     if verbose:
-        logger.info(str(G.shape[2])+' SNPs that pass filters')
+        # logger.info(str(G.shape[2])+' SNPs that pass filters')
+        print(str(G.shape[2])+' SNPs that pass filters')
     #### Fill NAs ####
     if verbose:
-        logger.info('Imputing missing values with population frequencies')
-    if not ignore_na_fams and not ignore_na_rows and not sib_diff and not fit_sib:
+        # logger.info('Imputing missing values with population frequencies')
+        print('Imputing missing values with population frequencies')
+    if not sib_diff and not fit_sib:
         # NAs = G.fill_NAs()
         # G = impute_missing(G)
         G.fill_NAs()
     elif sib_diff or fit_sib:
         G.fill_NAs()
-    
-    # if sib_diff:
-    #     print('1', G.gts[:, 0, :].mean())
-    #     print('doing extra stuff........')
-    #     G.gts[:, 0, :] -= G.gts[:, 1, :]
-    #     print('2', G.gts[:, 0, :].mean())
     if not robust:
         G.mean_normalise()
     ### Fit models for SNPs ###
-    # print(n_vararr)
-    # print("#"*16, min(varcomp_arr_lst[1][0])); exit()
     if n_vararr + 1 == len(varcomps):
-        model = lmm.LinearMixedModel(y, varcomp_arr_lst=varcomp_arr_lst,
+        model = slmm.LinearMixedModel(y, varcomp_arr_lst=varcomp_arr_lst,
                             varcomps=varcomps, covar_X=covar, add_intercept=True, add_jitter=add_jitter)
     elif n_vararr == len(varcomps) == 2: # grm is for imputation
-        print("#"*16, varcomps, min(varcomp_arr_lst[1][0]))
-        model = lmm.LinearMixedModel(y, varcomp_arr_lst=varcomp_arr_lst[1:],
+        model = slmm.LinearMixedModel(y, varcomp_arr_lst=varcomp_arr_lst[1:],
                             varcomps=varcomps, covar_X=covar, add_intercept=True, add_jitter=add_jitter)
     if robust:
-        # import pandas as pd
-        # pd.DataFrame({'ID': G.ids}).to_csv('../../output/robust_ids.txt', sep='\t', index=False)
-        # exit()
         alpha, alpha_cov, alpha_ses = model.robust_est(G.gts.data, G.num_obs_par_al, G.par_status)
     elif sib_diff:
-        # import pandas as pd
-        # pd.DataFrame({'ID': G.ids}).to_csv('sibdiff_ids.txt', sep='\t', index=False)
-        # exit()
-        alpha, alpha_cov, alpha_ses = model.sib_diff_est(G.gts.data, G.num_obs_par_al, G.par_status)
+        alpha, alpha_cov, alpha_ses = model.sib_diff_est(G.gts.data)
     else:
         if unrelated_inds is None or cond_gaussian:
-            alpha, alpha_cov, alpha_ses = model.fit_snps_eff(G.gts.data, G.fams, ignore_na_fams=ignore_na_fams, ignore_na_rows=ignore_na_rows, standard_gwas=standard_gwas)
+            alpha, alpha_cov, alpha_ses = model.fit_snps_eff(G.gts.data, standard_gwas=standard_gwas)
         else:
             alpha, alpha_cov, alpha_ses = model.fit_snps_eff_meta(G.gts.data, G.fams, unrelated_inds=unrelated_inds)
 
@@ -423,7 +314,7 @@ def process_batch(snp_ids, ped,  n_vararr, imp_fams, pheno_ids=None, bedfile=Non
 def process_chromosome(chrom_out, y, varcomp_lst,
                        ped, imp_fams, sigmas, outprefix, covariates=None, bedfile=None, bgenfile=None, par_gts_f=None,
                        fit_sib=False, sib_diff=False, parsum=False, impute_unrel=False, unrelated_inds=None, max_missing=5, min_maf=0.01, batch_size=10000, 
-                       no_hdf5_out=False, no_txt_out=False, cpus=1, debug=False, ignore_na_fams=False, ignore_na_rows=False, robust=False, cond_gaussian=False,
+                       no_hdf5_out=False, no_txt_out=False, cpus=1, debug=False, robust=False, cond_gaussian=False,
                        add_jitter=False, standard_gwas=False):
     ######## Check for bed/bgen #######
     if bedfile is None and bgenfile is None:
@@ -454,10 +345,12 @@ def process_chromosome(chrom_out, y, varcomp_lst,
         par_status, gt_indices, fam_labels = find_par_gts(y.ids, ped, gts_id_dict)
         parcount = np.sum(par_status==0,axis=1)
         if np.sum(parcount>0)==0:
-            logger.warning('No individuals with genotyped parents found. Using sum of imputed maternal and paternal genotypes to prevent collinearity.')
+            # logger.warning('No individuals with genotyped parents found. Using sum of imputed maternal and paternal genotypes to prevent collinearity.')
+            print('WARNING: no individuals with genotyped parents found. Using sum of imputed maternal and paternal genotypes to prevent collinearity.')
             parsum = True
         elif 100 > np.sum(parcount>0) > 0:
-            logger.warning('Warning: low number of individuals with observed parental genotypes. Consider using the --parsum argument to prevent issues due to collinearity.')
+            # logger.warning('Warning: low number of individuals with observed parental genotypes. Consider using the --parsum argument to prevent issues due to collinearity.')
+            print('WARNING: low number of individuals with observed parental genotypes. Consider using the --parsum argument to prevent issues due to collinearity.')
     ####### Compute batches #######
     print('Found '+str(snp_ids.shape[0])+' SNPs')
     # Remove duplicates
@@ -471,15 +364,8 @@ def process_chromosome(chrom_out, y, varcomp_lst,
         chrom = chrom[not_duplicated]
         alleles = alleles[not_duplicated,:]
     snp_dict = make_id_dict(snp_ids)
-    # Compute batches
-    # batch_bounds = compute_batch_boundaries(snp_ids,batch_size)
-    # if batch_bounds.shape[0] == 1:
-    #     print('Using 1 batch')
-    # else:
-    #     print('Using '+str(batch_bounds.shape[0])+' batches')
     if robust or sib_diff or standard_gwas:
         alpha_dim = 1
-        # if sib_diff: alpha_dim = 2
     else:
         alpha_dim = 2
         if fit_sib:
@@ -503,9 +389,11 @@ def process_chromosome(chrom_out, y, varcomp_lst,
     batches = split_batches(
         snp_ids.shape[0], len(y.ids), nbytes, cpus, parsum=parsum, sib=fit_sib or sib_diff)
     if len(batches) == 1:
-        logger.info('Using 1 batch')
+        # logger.info('Using 1 batch')
+        print('Using 1 batch')
     else:
-        logger.info('Using '+str(len(batches))+' batches')
+        # logger.info('Using '+str(len(batches))+' batches')
+        print('Using '+str(len(batches))+' batches')
     ##############  Process batches of SNPs ##############
     # make shared memory for multiprocessing
     y_ = RawArray('d', y.shape[0])
@@ -556,8 +444,6 @@ def process_chromosome(chrom_out, y, varcomp_lst,
                              max_missing=max_missing, min_maf=min_maf,
                              impute_unrel=impute_unrel,
                              unrelated_inds=unrelated_inds,
-                             ignore_na_fams=ignore_na_fams,
-                             ignore_na_rows=ignore_na_rows,
                              robust=robust,
                              cond_gaussian=cond_gaussian,
                              add_jitter=add_jitter)
@@ -565,9 +451,6 @@ def process_chromosome(chrom_out, y, varcomp_lst,
     if debug:
         _init_worker_(y_, len(varcomp_lst), varcomps_, covar_, covar_shape,) # ped_, ped_shape)
         batch_freqs, batch_snps, batch_alpha, batch_alpha_cov, batch_alpha_ses = process_batch_(snp_ids[batches[0]])
-        # gts = process_batch_(snp_ids[batches[0]])
-        # import code
-        # code.interact(local=locals())
         exit('Debug finishsed.')
     with Pool(
         processes=cpus,
@@ -577,48 +460,22 @@ def process_chromosome(chrom_out, y, varcomp_lst,
         result = pool.map(
             process_batch_, [snp_ids[ind] for ind in batches], chunksize=1)
     
-    ##### 删除
-    cnt = 0
-    # print('############')
-    # res = np.vstack(result)
-    # np.savez(f'accuracy_and_bias_{chrom_out}', res=res)
-    # print(np.quantile(res, [0, 0.25,0.5,0.75,1]))
-    # exit()
-        
-    #####
     for i in range(0, len(result)):
         batch_freqs, batch_snps, batch_alpha, batch_alpha_cov, batch_alpha_ses \
             = result[i]
         # Fill in fitted SNPs
         if len(batch_snps) == 0:
-            logger.info('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
+            # logger.info('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
+            print('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
             continue
         batch_indices = np.array([snp_dict[x] for x in batch_snps])
         alpha[batch_indices, :] = batch_alpha
         alpha_cov[batch_indices, :, :] = batch_alpha_cov
         alpha_ses[batch_indices, :] = batch_alpha_ses
         freqs[batch_indices] = batch_freqs
-        logger.info('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
-    ##############  Process batches of SNPs ##############
-    # for i in range(0,batch_bounds.shape[0]):
-    #     if i==0:
-    #         print_sample_info = True
-    #         verbose = True
-    #     else:
-    #         print_sample_info = False
-    #         verbose = False
-    #     batch_freqs, batch_snps, batch_alpha, batch_alpha_cov, batch_alpha_ses = process_batch(y, pedigree, 
-    #                 tau, sigma2, snp_ids=snp_ids[batch_bounds[i, 0]:batch_bounds[i, 1]], bedfile=bedfile, bgenfile=bgenfile, 
-    #                 par_gts_f = par_gts_f, parsum=parsum, fit_sib=fit_sib, max_missing=max_missing, min_maf=min_maf,
-    #                 print_sample_info=print_sample_info, verbose=verbose)
-    #     # Fill in fitted SNPs
-    #     batch_indices = np.array([snp_dict[x] for x in batch_snps])
-    #     alpha[batch_indices, :] = batch_alpha
-    #     alpha_cov[batch_indices, :, :] = batch_alpha_cov
-    #     alpha_ses[batch_indices, :] = batch_alpha_ses
-    #     freqs[batch_indices] = batch_freqs
-    #     print('Done batch '+str(i+1)+' out of '+str(batch_bounds.shape[0]))
-    ######## Save output #########
+        # logger.info('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
+        print('Done batch '+str(i+1)+' out of '+str(len(batches)) + f'#snps:{len(batch_snps)}')
+
     if not no_hdf5_out:
         if chrom_out==0:
             hdf5_outfile = outfile_name(outprefix, '.sumstats.hdf5')
