+++ snipar/read/__init__.py
@@ -135,10 +135,7 @@ def get_ids_with_par(gts_f: str,
         king['ID2'] = king['ID2'].astype(str)
 
         # remove one individual from each pair of MZ twins
-        # rm_mz = king[king['InfType'] == 'Dup/MZ'].apply(lambda row: random.choice([row.ID1, row.ID2]), axis=1).to_numpy()
         rm_mz = king[king['InfType'] == 'Dup/MZ'].apply(lambda row: row.ID1, axis=1).to_numpy()
-        # pd.DataFrame({'id': rm_mz}).to_csv('../../output/fastgwa/rm_mz.txt', sep='\t', index=False, header=None)
-        # exit()
         l = len(ids)
         if len(rm_mz) > 0:
             ids = np.setdiff1d(ids, rm_mz)
@@ -147,10 +144,6 @@ def get_ids_with_par(gts_f: str,
     if gts_f[(len(gts_f) - 4):len(gts_f)] == '.bed':
         gts_f_: Bed = Bed(gts_f, count_A1=True)
         gts_ids = gts_f_.iid[:, 1]
-        # logger.debug(f'Length of geno: {len(gts_ids)}')
-        # logger.debug('#iid' + len(gts_ids))
-        # logger.debug(f'#ids present in both phen and geno: {len(np.intersect1d(ids, gts_ids))}')
-        # logger.debug(f'#ids present in both ped and geno: {len(np.intersect1d(ped[:, 1], gts_ids))}')
         ids, observed_indices, imp_indices, parcount = preprocess.get_indices_given_ped(ped,                                
                                                                              gts_ids,
                                                                              imp_fams=imp_fams,
@@ -158,13 +151,9 @@ def get_ids_with_par(gts_f: str,
                                                                              sib=sib,
                                                                              include_unrel=include_unrel)
         gts_ids = gts_f_.iid[observed_indices, 1]
-        # logger.debug(f'Length of observed geno: {len(gts_ids)}')
     elif gts_f[(len(gts_f) - 5):len(gts_f)] == '.bgen':
         gts_f_ = open_bgen(gts_f)
         gts_ids = gts_f_.samples
-        # logger.debug(f'Length of geno: {len(gts_ids)}')
-        # logger.debug(f'#ids present in both phen and geno: {len(np.intersect1d(ids, gts_ids))}')
-        # logger.debug(f'#ids present in both ped and geno: {len(np.intersect1d(ped[:, 1], gts_ids))}')
         ids, observed_indices, imp_indices, parcount = preprocess.get_indices_given_ped(ped,
                                                                              gts_ids,
                                                                              imp_fams=imp_fams,
@@ -172,7 +161,6 @@ def get_ids_with_par(gts_f: str,
                                                                              sib=sib,
                                                                              include_unrel=include_unrel)
         gts_ids = gts_ids[observed_indices]
-        # logger.debug(f'Length of observed geno: {len(gts_ids)}')
     else:
         raise ValueError('Unknown filetype for observed genotypes file: ' +
                          str(gts_f))
@@ -183,82 +171,32 @@ def get_ids_with_par(gts_f: str,
 
     # Find indices in reduced data
     par_status, gt_indices, fam_labels = preprocess.find_par_gts(ids, ped, gts_id_dict, imp_fams)
-    # print(len(par_status), len(observed_indices))
-    # print(sum([i == j == -1 for i, j in par_status]))
-    # print(sum([i == j == -1 for _,i, j in gt_indices]))
-    # print(sum([1 for i in fam_labels if len(i) == 0]))
-    # ind = np.array([i for i in range(6797) if par_status[i, 0] == par_status[i, 1] == -1 and len(fam_labels[i]) != 0])
-    # print(par_status[ind], fam_labels[ind], gt_indices[ind])
-    # f = fam_labels[ind]
-    # print([ped[i, :] for i in range(len(ped)) if ped[i, 1] in f])
-
-    # logger.debug(f'Length of par status: {len(par_status)}')
-    # logger.debug(f'#missing [F, M]: {(par_status == -1).sum(axis=0)}')
-    # logger.debug(f'#missing F and M (from par_status): {sum([i == j == -1 for i, j in par_status])}')
-    # logger.debug(f'#missing F and M (fro gt_indices): {sum([i == j == -1 for _,i, j in gt_indices])}')
-    # logger.debug(f'#both par geno: {sum([i == j == 0 for _,i, j in gt_indices])}')
-    # logger.debug(f'#both par geno: {sum([i == j == 0 for i, j in par_status])}')
-    # logger.debug(f'#one par geno: {sum([i + j == 1 for i, j in par_status])}')
     n_empty_fams = sum([1 for i in fam_labels if len(i) == 0])
-    # logger.debug(f'#unique fam_labels: {np.unique(fam_labels).__len__()}')
-    # logger.debug(f'#empty fam_labels: {n_empty_fams}')
 
-    # logger.debug(fam_labels[fam_labels == ''])
-    # ind = fam_labels == ''
+
     fam_labels = fam_labels.astype('<U20')
     fam_labels[fam_labels == ''] = np.array([f'_not_{i}_' for i in range(n_empty_fams)], dtype='<U20')
-    # # ind = np.array([i[4] == 'False' and i[5] == 'False' for i in ped])
-    # logger.debug(f'#empty fam_labels after assignment: {sum([1 for i in fam_labels if len(i) == 0])}')
-    # logger.debug(f'#unique fam_labels: {np.unique(fam_labels).__len__()}')
-    # logger.debug(f'#iids to use {len(ids)}')
-    # xx = len([1 for i in fam_labels if '_not_' in i])
-    # logger.debug(f'{xx}')
 
     imp_obs_ids = [i for i in range(len(par_status)) if par_status[i, 0] >= 0 and par_status[i, 1] >= 0]
     imp_only_ids = [i for i in range(len(par_status)) if par_status[i, 0] == par_status[i, 1] == 1]
     obs_only_ids = [i for i in range(len(par_status)) if par_status[i, 0] == par_status[i, 1] == 0]
-    logger.info(f'{len(imp_obs_ids)} individuals with phenotype observations and complete observed/imputed genotype observations.')
-    logger.info(f'{len(imp_only_ids)} individuals with imputed but no observed parental genotypes.')
-    logger.info(f'{sum(par_status.sum(axis=1) == 1)} individuals with one imputed and one observed parental genotypes.')
-    logger.info(f'{len(obs_only_ids)} individuals with both parents observed.')
+    # logger.info(f'{len(imp_obs_ids)} individuals with phenotype observations and complete observed/imputed genotype observations.')
+    # logger.info(f'{len(imp_only_ids)} individuals with imputed but no observed parental genotypes.')
+    # logger.info(f'{sum(par_status.sum(axis=1) == 1)} individuals with one imputed and one observed parental genotypes.')
+    # logger.info(f'{len(obs_only_ids)} individuals with both parents observed.')
+    print(f'{len(imp_obs_ids)} individuals with phenotype observations and complete observed/imputed genotype observations.')
+    print(f'{len(imp_only_ids)} individuals with imputed but no observed parental genotypes.')
+    print(f'{sum(par_status.sum(axis=1) == 1)} individuals with one imputed and one observed parental genotypes.')
+    print(f'{len(obs_only_ids)} individuals with both parents observed.')
 
     if include_unrel:
-        # missing_ids = [i for i in range(len(par_status)) if par_status[i, 0] == -1 and par_status[i, 1] == -1]
         unrelated_inds = np.where(np.array([par_status[i, 0] == -1 and par_status[i, 1] == -1 for i in range(len(par_status))]))[0]
-        # logger.info(f'{len(missing_ids)} individuals with both parents missing.')
-        logger.info(f'{len(unrelated_inds)} individuals with both parents missing.')
+        # logger.info(f'{len(unrelated_inds)} individuals with both parents missing.')
+        print(f'{len(unrelated_inds)} individuals with both parents missing.')
     one_missing_ids = [i for i in range(len(par_status)) if sum(par_status[i, :]) <= 0 and par_status[i, 0] * par_status[i, 1] < 0]
     if len(one_missing_ids) > 0:
-        logger.error(f'{len(one_missing_ids)} individuals have one missing parent and one non-missing parent.')
+        raise RuntimeError(f'{len(one_missing_ids)} individuals have one missing parent and one non-missing parent.')
         exit(-1)
-    ##### testing code
-    # print(np.where(ped[:, 1] == ids[0])[0])
-    # ids_id = [np.where(ped[:, 1] == i)[0][0] for i in ids]
-    # np.testing.assert_array_equal(ped[ids_id, 0], fam_labels); print('ok', len(ids))
-    # print(len([1 for i,j in par_status if i >= 0 and j >= 0]), len(par_status))
-    # print(sum([i == j for _,i,j in gt_indices[imp_only_ids]]), len(imp_only_ids))
-    # print(sum([i != j for _,i,j in gt_indices[obs_only_ids]]), len(obs_only_ids))
-
-    # uni, unique_fam_inds, cnt = np.unique(fam_labels, return_index=True, return_counts=True)
-    # x = list(uni[:5])
-    # fam_ids = [i for i in range(len(fam_labels)) if fam_labels[i] in x]
-    # ids = ids[fam_ids]
-    # fam_labels = fam_labels[fam_ids]
-    nnz = 0
-    _, unique_fam_inds, cnt = np.unique(fam_labels, return_index=True, return_counts=True)
-    for i in cnt:
-        nnz += i * (i + 1) /2
-    logger.info(f'nnz should be {nnz}')
-
-    # unique_fam_ids = unique_fam_ids[cnt == 1]
-    # father_obs = [i for i in range(len(par_status)) if par_status[i, 0] == 0 and par_status[i, 1] == 1]
-    # father_obs = [i for i in range(len(par_status)) if par_status[i, 0] == 1 and par_status[i, 1] == 0]
-    # indices = np.intersect1d(father_obs, unique_fam_ids)
-    # indices = imp_only_ids
-    # ids = ids[indices]
-    # fam_labels = fam_labels[indices]
-    # ids_id = [np.where(ped[:, 1] == i)[0][0] for i in ids]
-    # print(sum([i == 'False' and j == 'False' for _, _, _, _, i, j in ped[ids_id, :]]), 'test if the ids truly have imputed father and observed mother')
     
     if return_info:
         if include_unrel:
@@ -268,7 +206,7 @@ def get_ids_with_par(gts_f: str,
 
 
 def get_ids_with_sibs(gts_f: str,
-                      par_gts_f: str,
+                      ped: np.ndarray, 
                       ids: np.ndarray = None,
                       ibdrel_path: str = None,
                       return_info: bool = False) -> Tuple[np.ndarray, np.ndarray]:
@@ -283,111 +221,44 @@ def get_ids_with_sibs(gts_f: str,
         # remove one individual from each pair of MZ twins
         # rm_mz = king[king['InfType'] == 'Dup/MZ'].apply(lambda row: random.choice([row.ID1, row.ID2]), axis=1).to_numpy()
         rm_mz = king[king['InfType'] == 'Dup/MZ'].apply(lambda row: row.ID1, axis=1).to_numpy()
+        l = len(ids)
         if len(rm_mz) > 0:
             ids = np.setdiff1d(ids, rm_mz)
-            print(f'WARNING: {len(rm_mz)} pairs of Duplicates or MZ twins in pedigree.')
-    # Imputed parental file
-    par_gts_f_ = h5py.File(par_gts_f, 'r')
-    # Genotype file
-    ped = convert_str_array(np.array(par_gts_f_['pedigree']))
-    ped = ped[1:ped.shape[0], :]
-    # logger.debug(f'ped {len(ped)}')
-    # Remove control families
-    controls = np.array([x[0] == '_' for x in ped[:, 0]])
-    ped = ped[np.logical_not(controls), :]
-    # logger.debug(f'ped {len(ped)}')
-    # Get families with imputed parental genotypes
-    fams = convert_str_array(np.array(par_gts_f_['families']))
-    # logger.debug(f'#ids present in both phen and ped: {len(np.intersect1d(ids, ped[:, 1]))}')
+            print(f'WARNING: {len(rm_mz)} pairs of Duplicates or MZ twins in pedigree. Removed {l - len(ids)} individuals.')
 
     if gts_f[(len(gts_f) - 4):len(gts_f)] == '.bed':
         gts_f_: Bed = Bed(gts_f, count_A1=True)
         gts_ids = gts_f_.iid[:, 1]
-        ids, observed_indices = preprocess.get_indices_given_ped_sibs(ped, gts_ids, imp_fams=fams, ids=ids, verbose=True)
+        ids, observed_indices = preprocess.get_indices_given_ped_sibs(ped, gts_ids, ids=ids, verbose=True)
         gts_ids = gts_f_.iid[observed_indices, 1]
     elif gts_f[(len(gts_f) - 5):len(gts_f)] == '.bgen':
         gts_f_ = open_bgen(gts_f)
         gts_ids = gts_f_.samples
-        ids, observed_indices = preprocess.get_indices_given_ped_sibs(ped, gts_ids, imp_fams=fams, ids=ids, verbose=True)
+        ids, observed_indices = preprocess.get_indices_given_ped_sibs(ped, gts_ids, ids=ids, verbose=True)
         gts_ids = gts_ids[observed_indices]
     else:
         raise ValueError('Unknown filetype for observed genotypes file: ' +
                          str(gts_f))
 
-    # fams = fams[imp_indices]
     gts_id_dict = make_id_dict(gts_ids)
 
     # Find indices in reduced data
-    gt_indices, fam_labels = preprocess.find_gts(ids, ped, gts_id_dict, fams)
+    gt_indices, fam_labels = preprocess.find_gts(ids, ped, gts_id_dict)
     # Find which individuals can be used
     none_missing = gt_indices >= 0
     N = np.sum(none_missing)
     if N == 0:
         raise ValueError(
-            'No individuals with phenotype observations and complete observed/imputed genotype observations')
+            'No individuals with phenotype observations and ')
     # Take those that can be used
-    gt_indices = gt_indices[none_missing]
+    # gt_indices = gt_indices[none_missing]
     ids = ids[none_missing]
-    # print(len(par_status), len(observed_indices))
-    # print(sum([i == j == -1 for i, j in par_status]))
-    # print(sum([i == j == -1 for _,i, j in gt_indices]))
-    # print(sum([1 for i in fam_labels if len(i) == 0]))
-    # ind = np.array([i for i in range(6797) if par_status[i, 0] == par_status[i, 1] == -1 and len(fam_labels[i]) != 0])
-    # print(par_status[ind], fam_labels[ind], gt_indices[ind])
-    # f = fam_labels[ind]
-    # print([ped[i, :] for i in range(len(ped)) if ped[i, 1] in f])
 
-    # logger.debug(f'Length of par status: {len(par_status)}')
-    # logger.debug(f'#missing [F, M]: {(par_status == -1).sum(axis=0)}')
-    # logger.debug(f'#missing F and M (from par_status): {sum([i == j == -1 for i, j in par_status])}')
-    # logger.debug(f'#missing F and M (fro gt_indices): {sum([i == j == -1 for _,i, j in gt_indices])}')
-    # logger.debug(f'#both par geno: {sum([i == j == 0 for _,i, j in gt_indices])}')
-    # logger.debug(f'#both par geno: {sum([i == j == 0 for i, j in par_status])}')
-    # logger.debug(f'#one par geno: {sum([i + j == 1 for i, j in par_status])}')
     n_empty_fams = sum([1 for i in fam_labels if len(i) == 0])
-    # logger.debug(f'#unique fam_labels: {np.unique(fam_labels).__len__()}')
-    # logger.debug(f'#empty fam_labels: {n_empty_fams}')
 
-    # logger.debug(fam_labels[fam_labels == ''])
-    # ind = fam_labels == ''
     fam_labels = fam_labels.astype('<U20')
     fam_labels[fam_labels == ''] = np.array([f'_not_{i}_' for i in range(n_empty_fams)], dtype='<U20')
-    # # ind = np.array([i[4] == 'False' and i[5] == 'False' for i in ped])
-    # logger.debug(f'#empty fam_labels after assignment: {sum([1 for i in fam_labels if len(i) == 0])}')
-    # logger.debug(f'#unique fam_labels: {np.unique(fam_labels).__len__()}')
-    # logger.debug(f'#iids to use {len(ids)}')
-    # xx = len([1 for i in fam_labels if '_not_' in i])
-    # logger.debug(f'{xx}')
 
-    ##### testing code
-    # print(np.where(ped[:, 1] == ids[0])[0])
-    # ids_id = [np.where(ped[:, 1] == i)[0][0] for i in ids]
-    # np.testing.assert_array_equal(ped[ids_id, 0], fam_labels); print('ok', len(ids))
-    # print(len([1 for i,j in par_status if i >= 0 and j >= 0]), len(par_status))
-    # print(sum([i == j for _,i,j in gt_indices[imp_only_ids]]), len(imp_only_ids))
-    # print(sum([i != j for _,i,j in gt_indices[obs_only_ids]]), len(obs_only_ids))
-
-    # uni, unique_fam_inds, cnt = np.unique(fam_labels, return_index=True, return_counts=True)
-    # x = list(uni[:5])
-    # fam_ids = [i for i in range(len(fam_labels)) if fam_labels[i] in x]
-    # ids = ids[fam_ids]
-    # fam_labels = fam_labels[fam_ids]
-    nnz = 0
-    _, unique_fam_inds, cnt = np.unique(fam_labels, return_index=True, return_counts=True)
-    for i in cnt:
-        nnz += i * (i + 1) /2
-    logger.info(f'nnz should be {nnz}')
-
-    # unique_fam_ids = unique_fam_ids[cnt == 1]
-    # father_obs = [i for i in range(len(par_status)) if par_status[i, 0] == 0 and par_status[i, 1] == 1]
-    # father_obs = [i for i in range(len(par_status)) if par_status[i, 0] == 1 and par_status[i, 1] == 0]
-    # indices = np.intersect1d(father_obs, unique_fam_ids)
-    # indices = imp_only_ids
-    # ids = ids[indices]
-    # fam_labels = fam_labels[indices]
-    # ids_id = [np.where(ped[:, 1] == i)[0][0] for i in ids]
-    # print(sum([i == 'False' and j == 'False' for _, _, _, _, i, j in ped[ids_id, :]]), 'test if the ids truly have imputed father and observed mother')
-    
     if return_info:
         return ids, fam_labels, gt_indices
     return ids, fam_labels
